+++ tests/testthat/test-16-getCellsTests.R
@@ -64,6 +64,40 @@ testScenarios <- function(description="test", releaseEvaluationMode="batch", rel
 context("GET CELLS TESTS")
 
 
+scenarios <- testScenarios("all cells test")
+for(i in 1:nrow(scenarios)) {
+  if(!isDevelopmentVersion) break
+  evaluationMode <- scenarios$evaluationMode[i]
+  processingLibrary <- scenarios$processingLibrary[i]
+  description <- scenarios$description[i]
+  countFunction <- scenarios$countFunction[i]
+
+  test_that(description, {
+
+    library(pivottabler)
+    pt <- PivotTable$new(processingLibrary=processingLibrary, evaluationMode=evaluationMode)
+    pt$addData(bhmtrains)
+    pt$addColumnDataGroups("TrainCategory")
+    pt$addRowDataGroups("TOC")
+    pt$defineCalculation(calculationName="TotalTrains", summariseExpression=countFunction)
+    pt$evaluatePivot()
+    # pt$renderPivot()
+    # sum(pt$cells$asMatrix(), na.rm=TRUE)
+
+    allCells <- pt$allCells
+
+    fx <- function(x) { is.null(x$rawValue) }
+    cellsWithValues <- allCells[!sapply(allCells, fx)]
+    fx <- function(x) { x$rawValue }
+    values <- sapply(cellsWithValues, fx)
+
+    expect_equal(sum(pt$cells$asMatrix(), na.rm=TRUE), 334840)
+    expect_equal(length(allCells), 15)
+    expect_equal(sum(Values), 334840)
+  })
+}
+
+
 scenarios <- testScenarios("get cell test")
 for(i in 1:nrow(scenarios)) {
   if(!isDevelopmentVersion) break